requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pymongo
bcrypt
pyjwt
pydantic
requests
numpy==1.26.4
wfdb
firebase-admin
protobuf==4.25.3
grpcio==1.62.2
grpcio-status==1.62.2
grpcio-tools==1.62.2
openai
pydantic[email]
resend